+++ bw_timex/__init__.py
@@ -10,7 +10,7 @@ from .edge_extractor import EdgeExtractor
 from .helper_classes import SetList
 from .matrix_modifier import MatrixModifier
 from .timeline_builder import TimelineBuilder
-from .timex_lca import TimexLCA
+from .timex_lca import TimexLCA, TimexLCASettings
 from .utils import (
     add_flows_to_characterization_functions,
     add_temporal_distribution_to_exchange,
@@ -30,6 +30,7 @@ __all__ = [
     "easy_timedelta_distribution",
     # core classes
     "TimexLCA",
+    "TimexLCASettings",
     "TimelineBuilder",
     "MatrixModifier",
     "DynamicBiosphereBuilder",